requirements.txt
asgiref==3.3.4
attrs==21.2.0
bson==0.5.10
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
dnspython==2.1.0
docker==5.0.0
fastapi==0.65.2
h11==0.12.0
idna==2.10
iniconfig==1.1.1
mangum==0.11.0
motor==2.4.0
packaging==20.9
pluggy==0.13.1
py==1.10.0
pydantic==1.8.2
pymongo==3.11.4
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
python-dotenv==0.18.0
requests==2.25.1
six==1.16.0
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.5
uvicorn==0.14.0
websocket-client==1.1.0
pytz==2021.1
ujson==4.0.2
opentracing==2.4.0
jaeger-client==4.5
threadloop==1.0.2
thrift==0.13.0
tornado==6.1
cffi==1.14.5
cryptography==3.4.7
httpcore==0.12.3
httpx==0.17.1
httpx-oauth==0.3.5
jwt==1.2.0
passlib==1.7.4
pycparser==2.20
rfc3986==1.5.0
sniffio==1.2.0
deprecated==1.2.12
makefun==1.11.3
wrapt==1.12.1
email-validator==1.1.3
python-multipart==0.0.5